feat: user_token_generation - vikas tomar

+++ batwara/authentication/views/authorization.py
@@ -6,6 +6,9 @@ from twilio.rest import Client
 from ..models import *
 from django.views.decorators.csrf import csrf_exempt
 from configuration import constants
+import jwt
+import datetime
+from datetime import timedelta
 
 # Function to generate a random OTP
 def generate_otp():
@@ -121,9 +124,26 @@ def opt_authentication(request):
         #To check the generated otp is same as user pass otp
         if not sent_user_otp == user_request['user_otp']:
             return JsonResponse({'data': 'Invalid OTP'},safe=False)
+        
+        user_token = generate_token(request)
+        if not user_token:
+            return JsonResponse({'data': 'Invalid OTP'},safe=False)
+            
         delete_otp_generated = TempOtp.objects.filter(user_phone=user_phone,otp_id=otp_id).delete()
-        return JsonResponse({'data': 'success','token':'temp_user_token'},safe=False,status=constants.HTTP_200_OK)
+        return JsonResponse({'data': 'success','token':user_token},safe=False,status=constants.HTTP_200_OK)
     except Exception as error:
         print(error)
         return JsonResponse({'data': 'error'},safe=False,status=constants.HTTP_500_INTERNAL_SERVER_ERROR)
+
+def generate_token(request):
+    try:
+        user_request = json.loads(request.body)
+        if user_request:
+            user_phone = user_request['user_phone']
+        expiry_time = datetime.datetime.now()+timedelta(minutes=constants.TOKEN_EXPIRY)
+        encoded_jwt = jwt.encode({"app_id": "batware", "exp": expiry_time}, user_phone, algorithm="HS256")
+        return encoded_jwt
+    except Exception as error:
+        print(error)
+        return False
